[ALPS03533148] spm: mt6775: early porting
[Detail]
-. introduce spm_v3 early porting
-. introduce gpio_dump_regs weak function
-. introduce vcorefs_get/eint print weak function
-. update spm_flags1/SPM_RSV_CON2 bit define
-. update spm_flags1 to atf flow

[Solution]
-. must build pass

MTK-Commit-Id: 28fa2429190508b9634c73d2c1f353e678c01d30

CR-Id: ALPS03533148
Feature: Power Management
(cherry picked from commit 15ddf99d0274c95b8ee85cda07dd2981a2ef1ff9)

+++ drivers/misc/mediatek/base/power/include/spm_v3/mtk_spm_internal.h
@@ -381,7 +381,7 @@ struct pwr_ctrl {
 #if defined(CONFIG_MACH_MT6759) \
 	|| defined(CONFIG_MACH_MT6758) \
 	|| defined(CONFIG_MACH_MT6775)
-	u8 spm_rsv_con2;
+	u32 spm_rsv_con2;
 #endif
 	/* Auto-gen End */
 };
